backend-codes-gd/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
django-environ==0.11.2
django-filter==23.5
celery==5.3.4
redis==5.0.1
channels==4.0.0
channels-redis==4.1.0
psycopg2-binary==2.9.9
Pillow==10.1.0
django-storages==1.14.2
boto3==1.34.0
django-crispy-forms==2.1
crispy-bootstrap5==0.7
django-allauth==0.57.0
stripe==7.8.0
paystackapi==2.1.0
requests==2.31.0
python-decouple==3.8
django-debug-toolbar==4.2.0
pytest-django==4.7.0
coverage==7.3.2
black==23.11.0
flake8==6.1.0
daphne==4.0.0
django-celery-beat==2.5.0
gunicorn==21.2.0
uvicorn==0.24.0
